Return all response headers in txn variables.
Headers are stored as variables in txn.auth_request_header.*, with
names being normalized to be easily accessible from HAProxy. A
list of the normalized return header names is also stored in
txn.auth_request_header_names to ease debugging.

As an example of normalization, a header named ".X-Example-Foo.Bar!"
will be normalized to "x_example_foo_bar" and will be stored in
the txn.auth_response_header.x_example_foo_bar variable.